requirements.txt
aiofiles==22.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aiosqlite==0.21.0
ansi2html==1.9.2
anyio==4.11.0
aquirdturtle-collapsible-headings==3.1.0
archspec==0.2.5
argo-workflows==3.6.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrdict==2.0.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.6.3
black==25.9.0
bleach==6.2.0
blinker==1.9.0
bokeh==3.8.0
boltons==24.0.0
Brotli==1.1.0
cachetools==6.2.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
comm==0.2.3
conda==25.1.0
conda-build==25.1.1
conda_index==0.5.0
conda-libmamba-solver==25.1.1
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
contourpy==1.3.3
csscompressor==0.9.5
cycler==0.12.1
daiquiri==3.4.0
dash==3.2.0
debugpy==1.8.17
decompyle3==3.9.4.dev0
decorator==5.1.1
defusedxml==0.7.1
delegator.py==0.1.1
dictdiffer==0.9.0
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
efficientnet_pytorch==0.7.1
einops==0.8.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
expecttest==0.3.0
fastjsonschema==2.21.2
filelock==3.17.0
Flask==3.1.2
fonttools==4.60.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
h11==0.16.0
h2==4.1.0
hf-xet==1.1.10
hpack==4.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.0.1
hypothesis==6.124.7
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
invectio==0.2.2
ipykernel==7.0.1
ipynbname==2025.8.0.0
ipython==9.6.0
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==7.0.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.5.2
json5==0.12.1
jsonformatter==0.3.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.9.0
jupyter-dash==0.4.2
jupyter-events==0.12.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.3.0
jupyter_nbextensions_configurator==0.6.4
jupyter-nbutils==0.1.3
jupyter-require==0.6.1
jupyter-resource-usage==0.7.2
jupyter_server==2.13.0
jupyter_server_fileid==0.9.3
jupyter-server-mathjax==0.2.6
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.8
jupyterlab_code_formatter==3.0.2
jupyterlab-drawio==0.9.0
jupyterlab_git==0.51.2
jupyterlab_latex==4.3.0
jupyterlab-logout==0.5.0
jupyterlab_materialdarker==0.7.1
jupyterlab-nvdashboard==0.9.0
jupyterlab_pygments==0.3.0
jupyterlab-quickopen==2.1.0
jupyterlab-requirements==0.16.2
jupyterlab_server==2.27.3
jupyterlab-system-monitor==0.8.0
jupyterlab-theme-hale==0.1.3
jupyterlab_theme_solarized_dark==3.0.1
jupyterlab-topbar==0.6.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
kubernetes==11.0.0
lazy_loader==0.4
lckr_jupyterlab_variableinspector==3.2.4
libarchive-c==5.1
libmambapy==2.0.5
lief==0.14.1
lime==0.2.0.1
lintrunner==0.12.7
llvmlite==0.45.1
lxml==6.0.2
mamba-ssm==2.2.6.post3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst==2.2.0
micropipenv==1.9.0
mistune==3.1.4
mock==5.2.0
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.8.0
nbclassic==1.3.3
nbclient==0.10.2
nbconvert==7.16.6
nbdime==4.0.2
nbformat==5.10.4
nbresuse==0.4.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
notebook==6.5.4
notebook_shim==0.2.4
numba==0.62.1
numpy==2.2.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==13.580.82
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.3.1
opencv-python==4.12.0.88
openshift==0.11.0
optree==0.14.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.3
pkginfo==1.12.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.3.1
pluggy==1.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.50
propcache==0.4.1
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat==0.6.6
pycparser==2.22
pyelftools==0.32
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-json-logger==4.0.0
python-string-utils==1.0.0
pytokens==0.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc5424-logging-handler==1.4.3
rich==14.2.0
rich-click==1.9.3
rpds-py==0.22.3
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
Send2Trash==1.8.3
sentry-sdk==2.42.0
setuptools==75.8.0
simpervisor==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spark-parser==1.9.0
stack_data==0.6.3
sympy==1.13.1
termcolor==3.1.0
terminado==0.18.1
thamos==1.29.1
theme-darcula==4.0.0
thoth-analyzer==0.1.8
thoth-common==0.36.6
thoth-python==0.17.0
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==1.0.20
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchelastic==0.2.2
torchvision==0.21.0+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.2.0
truststore==0.10.0
types-dataclasses==0.6.6
types-python-dateutil==2.9.0.20251008
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
umap==0.1.1
umap-learn==0.5.9.post2
uncompyle6==3.9.3
uri-template==1.3.0
urllib3==2.3.0
virtualenv==20.35.3
voluptuous==0.15.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
xdis==6.1.8
xeus-python==0.17.6
xeus-python-shell==0.6.4
xyzservices==2025.4.0
y-py==0.6.2
yarl==1.22.0
yaspin==3.3.0
ypy-websocket==0.8.4
zipp==3.21.0
zstandard==0.23.0
